separate causallm model test
