+++ .github/workflows/test_ipex.yml
@@ -23,7 +23,7 @@ jobs:
       matrix:
         transformers-version: ["4.49.0"]
         torch-version: ["2.6.0"]
-        test-file: [test_modeling.py, test_pipelines.py]
+        test-file: [test_modeling.py, test_pipelines.py, test_modeling_causal_lm.py]
 
     runs-on: ubuntu-22.04
 
@@ -43,14 +43,14 @@ jobs:
           pip install .[ipex,tests] transformers[testing]==${{ matrix.transformers-version }} intel_extension_for_pytorch==${{ matrix.torch-version }}
 
       - name: Install bitsandbytes
-        if: matrix.test-file == 'test_modeling.py'
+        if: matrix.test-file == 'test_modeling_causal_lm.py'
         run: |
           git clone --branch multi-backend-refactor https://github.com/bitsandbytes-foundation/bitsandbytes.git
           cd bitsandbytes
           pip install .
 
       - name: Install autoawq
-        if: matrix.test-file == 'test_modeling.py'
+        if: matrix.test-file == 'test_modeling_causal_lm.py'
         run: |
           git clone https://github.com/casper-hansen/AutoAWQ.git
           cd AutoAWQ
